custom_models/requirements.txt
openai-whisper
soundfile
uvicorn
fastapi
numpy
soundfile
torch
kokoro